Extract gem loader
