+++ lib/tapioca/internal.rb
@@ -27,7 +27,6 @@ require "yard-sorbet"
 
 require "tapioca/runtime/dynamic_mixin_compiler"
 require "tapioca/helpers/gem_helper"
-require "tapioca/runtime/loader"
 
 require "tapioca/helpers/sorbet_helper"
 require "tapioca/helpers/rbi_helper"
@@ -51,6 +50,7 @@ require "tapioca/static/symbol_loader"
 require "tapioca/static/requires_compiler"
 
 require "tapioca/loaders/loader"
+require "tapioca/loaders/gem"
 require "tapioca/loaders/dsl"
 
 require "tapioca/gem"
